Fix signed chat state de-sync caused by `execute run say`
`SignableCommand#of` goes down the command parse results to find signed
arguments which will form part of the signed chat state. This allows
commands like `/say` to have their messages be signed by the client.

When running through `/execute` (or any other similar command) however,
the message shouldn't be signed, as the command may be run by another
player, making the signature on the message meaningless.

To prevent this, `SignableCommand` stops searching when a command node
redirection loops back to the root command node (where it might go to a
command like `/say`). This allows other commands which use redirection
to still retain signed arguments, while disallowing `/execute` and
similar commands.

However, because of NeoForge's client commands system, the root command
node of the dispatcher and the root command node pointed by `/execute
run` for redirection are different, causing the stop check above to
never trigger, thus causing the client to sign the arguments.

As the server disregards the signatures, this now causes a de-sync
between the signed chat state of the client and the server; thus, the
client will be kicked by the server upon the next chat message, when it
realizes a de-sync happened.

To fix this, we generalize the check from being the original root
command node, to being *any* RootCommandNode. This retains compatibility
with the original vanilla check, while allowing for our custom root
command nodes to also count for the check.

Fixes #186