Rename certain DR subclass helpers as being simple
Certain helper methods on the two subclasses of `DeferredRegister` for
`Block`s and `Item`s offer convenience by using the normal `Block`,
`Item`, and `BlockItem` classes instead of a more specific type. They're
intended to be used in simple cases where no custom type is needed.

Renaming these methods to include `simple` in their name reduces
possible confusion that may arise. This is particularly evident with
the previously-named `registerBlock` overload. A user may use that
method while intending to pass a `Supplier`, not realizing the proper
method would be the `register` method, and facing a compile error about
their lambda's target (which is `BlockBehaviour.Properties`) not being a
functional interface.