[1.20.4] Fixed NeoForge pack version and description (Attempt 2)
Truly fixes these for real this timey
Fixes #278
Fixes #360

This is a **breaking change** because we are removing the obsolete server/resource pack numbers because vanilla allows us to specify version ranges. mods can just say they can run on all versions and avoid the red highlighting in pack screens entirely.

Thus no need for many of the patches here.