Make player feet and eyes both need to be in fluid to swim
Fixes https://github.com/neoforged/NeoForge/issues/220

Issue was the patch to Entity$updateSwimming removed the vanilla check that verified the player's feet was in water before allowing swimming. I added back the feet check for canSwim fluidtype.

The vanilla behavior in my testing is both the eye and feet bottom has to be in the fluid for swimming to be allowed. If either are out of fluid, you cannot go into swimming mode.

In testing this fix, I get that exact behavior now with neoforge. Both eyes and feet has to be in fluid to allow swimming which makes sense as canStartSwimming already checks eyes and now my check adds back feet