[1.20.4] Reimplement ID offset on modded EntityDataSerializers (#425)
This reimplements the ID offset previously applied to modded
`EntityDataSerializer`s which was lost during the registry rework.

Previously this was achieved by setting a minimum ID on the registry.
Due to the modified vanilla registries not supporting that, the IDs of
modded serializers now also start at 0 which leads to the client using
the vanilla serializer corresponding to the ID of the modded one and
potentially blowing up while deserializing the data item, i.e. by
reading past the buffer's length.