NeoForge chunk generate command improvements (#452)
Based on some additional feedback from:
https://github.com/jaskarth/fabric-chunkpregenerator/pull/27

I removed the workaround fix radius check out of GenerationTask class
and instead, added the missing x radius check into CoarseOnionIterator
since that was the real original cause of the command going beyond 100%
at times.

I also removed scheduling the completable future in a server task
runnable because that was an unneeded redirection by mistake. I also
added comments to state which parts of the code is safe to be on off
thread and which has to be on server to help future maintainability.
Added correct executor to the runAsync call as well