Load en_us languages on dedicated server
Ordinarily, Minecraft loads its English language asset file on the
server as a default (and sole) language, to allow translated messages
to be readable in the server console, such as those outputted by
commands.

While NeoForge also does this for itself, the code that loads the
language files from all other loaded resource packs is broken, because
`LanguageHook#loadLanguage` uses the datapack resource manager -- thus
it looks for `data/<namespace>/lang/en_us.json`.

The simplest and most correct fix is to create a throwaway resource
manager using the same pack list, but configured for client assets
(through the PackType), and using that to find and load the language
files.