Removed usage of static ModLoadingContext, used supplied modId instead
For this solution to work, the verification that a registered payload's namespace needs to match the mod id of the registrar had to be removed, in order to allow mods passing their proper mod id to the registrar while still being able to register payloads for any namespace they desire using that registrar (as has been one of the ideas of the Network Refactor).
Suggestion: put a duplication check in NetworkRegistry#setup where the channels are built, since currently duplicate payload ids just seem to be silently overwriting each other, which doesn't feel intentional.