Hoist the `Block#playerWillDestroy` call out of `IBlockExtension#onDestroyedByPlayer` (#1034)
Rewrites our patch logic for block break, moving the `Block#playerWillDestroy` call to the original call site, instead of having it be in our extension method.  
This change resolves an issue where Decorative Pots would drop themselves instead of Bricks/Sherds when broken with tools like Swords.  
Also propagates the result of `Block#playerWillDestroy` to `Block#destroy` on the client side (vanilla only does it on the server side).