Fixes for the Config UI
(a) Fix integer overflow when checking range of integers that allow more than 31 bits of range
(b) Change how sections and lists are shown, moving the name to the label and a generic "Edit..." onto the button. The button text can be overridden by translation for each individual button
(c) Fixed numeric configuration keys without ranges not being recognised
(d) Added simple value/section filter as that seems to be the most common reason for subclassing
(e) Updated tests with cases for all of a-d
(f) Changed element to include the "undoable" flag as the old logic exploded due to c

Technically breaking, but all changes were in places mods should not call directly (context and element default constructor).