Ensure that DummySavedData doesn't attempt to save
There is a potential race condition/edge case during crashes on startup which
may cause data to be overwritten:
- DimensionDataStorage#get is called
  - We patch this method to save a "dummy" instance in the cache
    when readSavedData returns null
- An error happens elsewhere, not necessarily directly related to saving/loading
- DimensionDataStorage#readSavedData returns null
  - This could be due to an exception, thread interrupt, etc - who knows!
  - By design, on an exception it returns null and swallows the error
    (it is, at the very least, logged)
- The dummy instance is saved into the cache
- DimensionDataStorage#save is called because the server is crashing
  - This iterates over the members of the cache and saves them
- DummySavedData#save is called
  - This **overwrites** any data that was previously in the file!

So, to fix this we simply override the DummySavedData to not even attempt
to write to the file. While this *does* mean that a corrupt file will
fail to load every time on server startup and not get cleared, data will
never be lost in this potential scenario.