Use a no-stacktrace exception for mod loading crash reports
The stacktrace from the 'root' exception of a mod loading crash report
is unnecessary since the crash report method is only called from a
limited number of places, which is what the stacktrace includes.

Removing the stacktrace makes the crash report more readable, by
removing a large part of it from the start that is irrelevant to actual
mod loading issues.