Translate failure message when possible
To make the crash reports more readable, the failure message is
translated from its original translation key to the English locale.
(English is used since crash reports are non-localized, and likely to be
shared with English-speaking developers.)

If the translation fails (such as due to an exception in an argument
formatter), then fallback to the previous behavior of adding the
translation key and its arguments separately.