Increase memory used for forked compiler
When running the Gradle Daemon with a non-J21 JVM, it will fork the
Java compiler into a separate process rather than running it in-process
with the Daemon. However, that means the compiler process uses the
default amount of memory, which is not able to compile all of NeoForge &
Minecraft (as it fails with an OutOfMemoryError).

Setting the memory when the compiler is forked ensures that whether the
compiler runs in-process with the Daemon or as a forked process, it will
always have enough memory to compile all the sources properly.