Fix index in `ItemHandlerCopySlot#setStackCopy`
The current code is using `index`, which is `this.index`, which ends up as `Slot#index`. That is not correct as that index is the index of the slot inside the menu, not inside the item handler. The correct index to use is that of `SlotItemHandler` (the name is the same as the one in the root `Slot`, but the class has access to it as it's protected and we're accessing across the same package)