Only fire gather data events for the mods enabled
Previously, all mods would receive the event and register to the data generator and the cli args will simply decide which generator to run. This can cause issues when dependencies ship their datagen which requires other dependencies as you'll hit CNFEs.

This PR changes it so that only the requested mods get the event posted.