Add a biome villagers data map to replace `VillagerType#BY_BIOME`
This PR adds a `neoforge:biome_villagers` biome data map to replace `VillagerType#BY_BIOME`, allowing mods to change the type of the villagers spawned in their own biomes.

To avoid a sudden breaking change, the map will still be used for now, but the fallback to it will be removed in 1.22.