Adds Notes in ADRIAN&SOARYN.txt which will be deleted by the time this PR work is over along with a lot of the note based comments
Organizes the IResourceHandler methods to help during implementation time.

While Adrian has done a fantastic job documenting, I added a few more points to some of the methods to try to clarify their intent. Especially on allowsInsert/Extraction as they are "stateless" and should be considered constant

Renames VoidHandler to be VoidResourceHandler just to avoid a small portion of confusion. Verbosity is fine here.

Commented out the uses of allowsExtraction/Insertion inside the actual behaviours. These methods are hints to things like pipe graph logistics or similar, not logic controllers. We can discuss an alternate approach to those if it is still feeling necessary to have something in place there. As I have a solution separately in my own implementations of IResourceHandler in XyCraft

Removes some abstraction from ISingleResourceHandler to avoid possible confusions as well a bit of clutter in terms of the number of methods that got populated from the abstraction. Ignoring parameters is fine, but we don't need to add a method just for wiping out the parameters.

Finalized the EmptyHandler (turning it into a record in this case). At first I moved these to just Anonymous classes in HandlerUtil, but saw the benefit of just being able to rapidly make a new instance for a brand new Resource Type. Changes the empty to return size 0.

Adds a IResourceStack to allow immutable and mutable resource stack to share properties. The idea behind the MutableResourceStack is to have all of the benefits of an immutable resource, but allowing a mutable count. This is needed in scenarios of ResourceStorageContents where the number will constantly be in flux. Instead of allocating constantly, we can focus on allocating only when we change the backing resource instead of every operation. (The old ResourceStorageContents is present has been labeled Deprecated and renamed Old* in case there was something we should use from that.)

In regards to MutableResourceStack, we can definitely look into cleaning it up, but it would preferred if we were to keep this class or have some ResourceIndex like thing.