Refactors structure of pr. This should help devs locate what they are looking for a little faster when trying to reference the source code.
Fully fixes all compiler errors associated with PR.

Adds an access transformer for DefaultDispenseItemBehavior to help with sound and animation.

Removes AggregateResourceHandler and replaces it with CombinedResourceWrapper. There is a problem at the moment with the combined when using it with Dynamically sized handlers. This will either be a different wrapper in the future or implemented on this variant. Neither of the scenarios, a handler reporting a changing size, nor is caching the size preemptively is incorrect; but it needs to be handled on a per use which is why we will likely have another wrapper handling the dynamic size.

Adds GameTests to handle a wide variety of scenarios as well as validates old tests there were present.

Temporary adds a group test for proof of concept but this should likely be removed by the time the pr is done. Feedback for Maty was given to see if we can get the results to print out in groups based on the `TestGroup`

Fixes HandlerIndexWrapper to have size of `1`. Otherwise it will assume the size of the parent.

Documents upgrade paths for IFluidHandler & IItemHandler in their respective methods, as well as in handler.md

Added the ability for the (we'll call them "Adrian templates") to be near allocation free when writing to an attachment, and allocates minimally when making an DataComponent to store on something like an ItemStack. Tests in `ComponentResourceTests` show that this is working as expected.

Pivoted back to having a `ISingleResourceHandler` given feedback of desires of an "index-less" storage.

This is something to discuss, but I've added the ResourceContainer that is used in XyCraft. The Item variant is the one that has been working in production for 2 years, but it has been reworked a while back to support IResource. This is not critical to leave, but it has made testing as well as explaining a simple implementation very easy. See TestResourceContainerAttachment for an example of implementations and ResourceHandlerUtilTests for consumption. Having both example templates (The ResourceContainer and ResourceStorage) might be fine, but we may want to look into what we need from both of them and isolate what is the best path if both are deemed unnecessary.

Removes optional calls from calls that should either be a resource or empty.

Merged similar codec & stream codec creations into IResourceStack for consistent data layouts / future refactoring