Documents some of the tests as to what they are expected to do for the next person who ever looks at them in the future to not be like "What... is this?"
Moved ISingleResourceHandler out of templates and more first class use those who wish to "Not worry about indices".

Refactors Handlers into a specific folder given there is an energy one in planning as well as allows for other "transfer" like types to be used though not sure that there is much need for that. A clearer distinction on what is the handler, template, and util. Utils in this sense should go out in transfer.

Adds a hashCode to IResourceStack helper for getting a hashcode from a list of the resourcestacks, like vanilla has for ItemStacks.