Moves the ItemHandlers to the respective transfer package.
Adds a utility method for extracting at a specific index

Fixes immutability of data components in Resource types. Someone could grab the map and mutate it, so now we return an immutable map.

Fixes default resources to actually have correct data on them. Creating them in the constructor is too early, and having them not be final is non-idealistic. Instead, a lazy is created in constructor and now lazily evaluated on demand. These are invalidated on default components modification. Fluids have the same notion, but there is no modification event.