Fix damage-blocking items not working
The incorrect order of `DamageContainer#setBlockedDamage` and
`LivingShieldBlockEvent#getBlockedDamage` causes the latter to return an
incorrect value since its presumption that the container still holds the
original damage is broken.

Switch the order of the two calls so the blocked damage is stored first
from the event before the damage container is updated.

Fixes #2066