Make any holdersets always serialize and the not one delegate the check
This PR makes the `AnyHolderSet` always return `true` (since the check is quite useless and it is possible to get technically valid registry lookups that are not of the same instance as the one passed in `canSerializeIn` - see #2073) and the `NotHolderSet` delegate the check to its `value` to fix #2073

Additionally, in order to fix #1838 `RegistrySetBuilder$UniversalLookup` is patched to implement the remaining trivial methods of `RegistryLookup`. This means that a `RegistryLookup` for a dynamic registry can be acquired from a registry bootstrapper `BootstrapContext`.