Use `IItemHandler#getSlotLimit` in `SlotItemHandler#getMaxStackSize`
The current implementation of `SlotItemHandler#getMaxStackSize(ItemStack)` attempts to insert the stack's max size into the handler and based on the current count in the slot, and the remainder of that operation, calculates the max stack size.
However, as pointed out in #1896 this causes needless slot modifications which can have unexpected side effects. Therefore this PR changes it to use the minimum between `IItemHandler#getSlotLimit`, and the item's max stack size.

As for *why* the current code looks like this, it seems like it was changed to swap the stack in the slot while calculating in https://github.com/neoforged/NeoForge/commit/a1bbcf8a0b840df9af1b01163d12af36eaf972bf, a whopping 9 years ago, when `IItemHandler` did not have a `#getSlotLimit` method. Ironically, just a month later, in https://github.com/neoforged/NeoForge/commit/c17b40790b616a59277ca00cff97fbb336c1369e, `IItemHandler#getSlotLimit` was introduced but `SlotItemHandler` was not modified to use it.

Fixes #1896