Adds a NotificationSnapshot to use a custom instanced EmptyValue instead of a bool. Functionally it should be the same, but this may serve as a slightly clearer pattern for those who want to extend it and make a OnChangedSnapshot or similar structures.
Reverts changes to snapshot bookkeeper regarding the int2obj map. This was profiled to be slower just by a TEEEENSY amount, so while not really worrisome, since we know of a better solution and is simplistic to use, we should use that. The while loop just looks odd, but that is just the nature of how java collections work.

Refactors documentation to make it simpler to read when not in an IDE.