Still removing dead comments.
Inlined IRegisteredResource, as the original intent was to make helpers and utilities on them. That is unnecessary to the bin it goes!

Moved ResourceContainer and related classes to a remove package for if we are given the goahead on merging this pr. The idea is to maintain the ability to compile and utilize, so until then we want to ensure they can work with any changes that may come from feedback. ALso added a readme for that package incase it is needed.

Renames generics to use Single characters instead words. This as I was told is more inline with Neo's normal design, so I abbreviated them with the first significant letter of the name instead.