On per permission from Orion, this adds energy handling from the sibling PR I was originally working on to go after this PR. This specifically targets making all the handlers be consistent with each other in names and use to some degree. Provided is a migration table in energy_migration.md as well as a rough entry primer energy_primer.md to help with review/understanding. This is a separate interface from the resource one `IEnergyHandler`.
Removes special onVaporized and similar methods in FluidResource. We were losing the amount info which may be relevant on what is played in those callbacks. We also couldn't safely pass the fluidstack with unsafe access as there would be no guarantee the fluidstack wouldn't be mutated in those calls. So instead we allocate with toStack for safety. Since it is a FluidBlockHandler, it should be fine.

Adds a limit to the snapshot depth storage post usage. The number we set here is currently arbitrary, though we should discuss what default value would be likely ideal, as well as the possibility of just making it an overridable amount per snapshot journal. This limit ONLY is for snapshot post resolution. So we aren't clamping the snapshot list when being added to, only resizing down to that size after committing/aborting.

Has a few extra classes in at the moment with regard to transactions. There was a design scope that Nano and Orion were discussing that we were playing around with as a possibility for utils. However, they are currently non-functional and not necessarily possible without capturing lambdas the way we were hoping to use them. It is something we were going to look into after reconvening. Lives got busy, so bear with us.

Adds a rough primer for IResources and IResourceHandlers. Expect this to be different and more informative later, but for the moment it should hopefully help a little.

Fixes insert/extract on WorldlyContainer. Somehow these were completely wiped, which may have been when merging onto the branch initially.

Changes `matches(stack)` to `is(stack)` in resource implementations