Changes BucketResourceHandler to return true for supports. While we could make this be set based on the filled value, as those should always be a different capability anyway; we are going to err on the side of caution to just return true.
Changes id of energy handler token from "energy" to "energy_handler"

Renames InventoryWrapper to PlayerInventoryWrapper

internalizes WrapperLocation to the cauldron and composter. This could have had unintended side effects of memory not being released if used by 3rd party by accident.

Adds null checks to ContainerOrHandler. We may want to investigate removing that class sometime as I get constantly told about abstraction...

Fixed potential issue if vanilla were to change behaviours in DropperBlock per pup's recommendation.

Adds back many of the old Utilities but with deprecations. Where we can, there is a bouncer with a migration path. Otherwise, they are intended to throw, and we provide an optimistic migration path in javadoc. Leaving FluidUtil 100% intact is not possible, nor should we really try. This is being introduced in a breaking change port.

Commented out FluidHandlerWrapper for now as the intent of it is, while admirable, not necessarily viable.

Fixed toStack in FluidResource to ensure we return empty with a `0` amount

Migrates final `@Range` call.

Refactors IItemContexts that return some modification of itself to their own file, while leaving in the default instancing method.

Still working on more Util methods