Removes `IResourceHandlerModifiable`. `.set()` is now inlined into each class and is public in the case of wanting to use say the templates for a ResourceHandlerSlot or similar.
Discovers issues with tests that will be worked on tomorrow or the next day to replace use of the wrong templates. ResourceContainers were used to test the infrastructure, but they should be swapped out for the other templates to test function as well given the ResourceContainers will be removed soon (either this week or next).