Adds public codec key to StackListHandler
Removes inherited java docs from ISingleEnergyHandler. The documentation is in IEnergyHandler.

Moves indexless calls for getAmount and getCapacity to IEnergyHandler defaulted on getAmount on the indices. Also does the same swap behaviour in ISingleEnergyHandler as ISingleResourceHandler does where the indexless one is the one recommended to override in handlers sized 1. This should serve as a decent compromise between those who don't want indices at ALL to those who do.