Moves the ItemContextResourceHandler to use a simpler DataComponent that only stores a single ResourceStack of some resource type. The capacity is now a hardcoded value rather than stored on the component.
 Applies IntMath.saturateMutliply to various returns that was using itemcontext results to clamp to MaxInt.

Simplifies the notification snapshot classes to a single GroupedSnapshotJournal which helps provide a way to signal when something commits or reverts instead of in the usual case scenario over multiple indices calling setChanged over and over, we can just call it once. In short, this is a helper snapshot that is responsible for calling setChanged once for an entire index snapshot chain.

 Removes unneeded snapshotting in the item component handlers, given they are in some container that is expected to handle the snapshot.

 Changes how the hashcode is generated in ResourceStack