Adds a safety for the UnsafeTransactionManager.getCurrentOpenedTransaction to return null if depth is -1
Uses the "current" transaction in simulated check methods in EnergyHandlerUtil & ResourceHandlerUtil. These are exclusive to transactions that would be non-committed.

Inlines TransactionContext.ROOT null value to all root opening calls.

Adds defaults to index-less operations in handlers, but still advises to handle with care when going to implement them.

Migrates some tests to JUnit. Was expecting this to be faster, but it takes a while for it to actually start the tests.