More later today. Need a break though for an hour or two.
Fixes various component related bugs, and adds a test for the ItemContainerContentsResourceHandler using shulker boxes in ComponentResourceTests.

getAmount helpers in ResourceHandlerUtils fixed to check the resource rather than skip the sought resource.

Starts rebuilding the template for the resource component based handler. Essentially the ItemContainerContents, but resource based. This will replace the ResourceStorageComponentHandler as that was designed differently with shared logic between two semi incompatible systems. Since the sibling system of that is gone, there is no reason for it to be structured the way it is now.

Changed the `with` in ItemContainerContents.java.patch as it would never have the correct size since something like shulkerboxes start at a trimmed size.