More pup reviews
Fixes some scenarios a transaction could be opened inside another without being aware in utils.

Moves component storage contents to new ResourceContainerContents & its handler. They are not documented yet, but the idea is structurally similar to the ItemContainerContents, but allows for resources to be held instead.

Starts removal of MutableResourceStack. Inlining the IResourceStack comes next.