Adds some basic tests for resource stacks and resources. Mostly validating the resource stack creation methods.
Adds an overload for `with` method in ResourceStack that takes in an UnaryOperator and an amount.

Removes construction method from ResourceStack and replaces it with a null check on the Resource's instanced empty stack. A single null check is inexpensive in this regard. There can be cases where an empty resource is valid info, and shouldn't be lost to a `?` implementation.