Makes TransferPreconditions.checkNonNegative return void.
Removes toExpandedString methods & privatizes optional methods we should leave them in though, for codec clarity and readability.

Moves TestElementResource from other slice to this slice. We want to be able to support these forms of resources (enums, records, etc.). This was one of the targets for the original PR for modded resources to be able to do.