Add method to wrap gui layers
This PR adds a new `RegisterGuiLayersEvent#wrapLayer` method which can be used to wrap a gui layer with an unary operator, allowing mods to, for instance, translate the location of the layer, thereby closing #2214.

I am however aware of the fact that the same can be achieved by listening to both the `RenderGuiLayerEvent.Pre` and the `RenderGuiLayerEvent.Post` events in order to correctly match the pose stacks. I do think though, that doing so is more harder and requires more care to ensure there is no mismatch present, so that's why I'm making this PR. If others feel like there is no need for these methods, feel free to comment as such.