Add `CompoundTag`/`ListTag` constructors with initial size
Adds a `ListTag` constructor with an int parameter specifying the initial size of the array list so that it does not need to be resized when the amount of expected elements is known.
A similar `CompoundTag` constructor is also added, though it uses `HashMap#newHashMap` which accounts for the default 0.75 map load factor.