Removes lazily initialized filled bucket in FluidResource. In theory allocation will be seldom and cheap enough it won't cause issues.
Adds missing `this` in Item constructor for the default resource. Functionally changes nothing, but it should be consistent.

Changes name of invalidateDefault -> createDefaultInstance to better indicate it is used to create the instance in the Item class when construction happens as well as modifying default components.

Removes isEnabled from IRegisteredResource and resource stack.

Fixes mistaken use of constructor in `with` patch calls in built in resources. Should use the `of` method to ensure we use the default resource instance where possible.