Fix MC-219981 by healing leader zombie health relative to new max health
[MC-219981](https://bugs.mojang.com/browse/MC/issues/MC-219981) is a confirmed vanilla bug which makes leader zombies not heal relative to their max health after becoming a leader (either by natural spawning or conversion to husk or drowned).
To fix it, we re-set the health of the zombie after turning into a leader, but we do so taking into consideration both its new max health and the amount of damage it had previously taken. Therefore a 20/20 zombie turning into a leader with a max health of 48 will have 48 health while a 15/20 zombie turning into a leader with a max health of 60 will have 55/60 health.
Fixes #2095