Add `ContextOps` as a general stackable abstraction for context-aware dynamic ops
Adds a `ContextOps` that can hold arbitrary context objects identified by a `Key` (a `Class` type and a `ResourceLocation` identifier).

The motivation for this is that currently, in order to inject additional context into dynamic ops for use in codecs, one has to create a concrete chain of `DelegatingOps` child classes, knowing exactly what context is provided.

This system is also intended to be used in the future to inject the `AttachmentHolder` into holder-aware but codec-backed data attachments. Another possible use case is a future event that allows mods to inject custom context into vanilla-deserialised objects like datapack registries, recipes, advancements, loot tables, etc. Obviously this system can also be immediately used by mods in their own codec-based deserialisation code.

Unfortunately, the vanilla `RegistryOps` cannot be feasibly patched out to a `ContextOps.Key` due to the amount of references to the class, so it is expected that this ops (which now extends from `ContextOps` itself) sits at the base of any context chain that makes use of registry lookups. The NeoForge-added `ConditionalOps` has however been modified to be backed by context injection. To preserve backwards compatibility, for now the class is still an ops instance. The `ConditionalOps.retrieveContext` method does not on the other hand pose the requirement of a `ConditionalOps` instance, relying instead on context injection.

Tests will be added shortly.