Changes migration debugging info into SnapshotJournalDebugInfo to present a little more information, as well as make it cleaner/simpler to remove post migration.
Removes callback methods from TransactionContext. This all should now be handled by SnapshotJournals on the consumer side should they need more handling. Still may be things we can clean up about this.

Moved `Result` to Transaction and made it private/

Moves static methods from TransactionManager to Transaction to match style of things like `List.of`

Removes unnecessary Internal annotations on package private members

Adds a StatelessSnapshot base that is equivalent in use to `SnapshotJournal<Void>` while keeping the infrastructure the same. An attempt at making SnapshotJournal<Void> was made.

Changes name of `nestingDepth` to just `depth` as it was redundant. In this case depth is always nested in the context of a transaction.

Handles wording of documentation to try to keep it consistent. There are still places to improve intent.

Adds a note about the UnsafeTransactionManager deprecation. It is intended to be removed, but inlining is only possible when the slices are available, and I'd rather not waste time manually doing that.