Changes `isActive` to be an instance method in TransactionContext, and renames the static method to `hasActiveTransaction` as well as adds a `lifecycle()` instanced getter method.
Renames `closeableJournals` to `journalsToCommit` to more clearly express these journals are marked as what should commit for a given depth.

Fixes some comment consistency with spaces after the `//`

Removes public access to getOpenTransaction with an arbitrary depth. Users should not be accessing other transactions that they could inadvertently commit. Instead, the only transactions they should have are the contexts or the ones they opened.

Reworks the close callbacks a little more. To make it more simplified down to just the transaction that can interact with it. This means that the only place you can interact with adding these callbacks or change what would happen during these times reside in SnapshotJournal.