Moves caller source class to a field on Transaction rather than an int map on TransactionManager. We leave it as a class since we only need to resolve it as a string when it would crash in specific scenarios.
Removes annotation for test order. Originally it was there as intellij was seemingly having problems picking up it was a test, but this seems to have resolved itself at some point. (It is highly possible it was still indexing when I was first writing it)

Removes snapshot templates to look at in a later slice.

Removes transaction debug metric info due to concerns of thread safety; however, we should likely look into something at some point that we can use to debug this information quickly, to validate our expectations of how large a transaction depth can get.

Changes cast operation to a direct cast in SnapshotJournal.