Seals `TransactionContext` permitting only `Transaction` to prevent custom implementations from being used. These would have otherwise crashed when used with a normal flow of the transaction system.
Renames SnapshotJournal's `onClose` to `close` and makes it package private. Since this is no longer a user facing callback, it should be more aligned with being a directive.

Removes `OverrideOnly` annotations for methods in SnapshotJournal.

Adds two new tests for simultaneous transaction validations one for root, one for shared parent.