Changes `ResourceStack` to a record. This will cause issues in use later and push the burdens onto the user, but we are to "see how modders use it and find out".
Removes EmptyResourceInfo with the removal of ResourceStack utility. This does cause things like the codecs to now need a stack factory since we can't gather instanced fields anymore.

Adds safeguards to prevent ResourceStacks from having a resource and a zero amount or an empty resource and a non-zero amount.
This should mean the only 2 valid instances are fully empty, or quantified with a positive value.

Removes specific readonly methods such as canEquip for Item/Fluid resources for a later slice as requested.