Add tool stack sensitive overload of `onDestroyedByPlayer`
On the server, by the time `onDestroyedByPlayer` has been called, the players tool stack (main hand) has already been damaged, and may be empty.

Currently to get the tool stack for custom drops (such as destroying part of the block) in this method, you need to ThreadLocal store it from `canHarvestBlock` which is less than ideal.