Fix fluid-pushing being performed twice
`LivingEntity#checkFallDamage` accidentally causes a second fluid-push
to be done, when it's meant only to double-check if the entity is in
water.

Modify `Entity#updateFluidHeightAndDoFluidPushing` to allow calling it
to update fluid height without performing fluid pushing, and update
other call-sites accordingly to ensure `LivingEntity#checkFallDamage`
makes use of the parameter to avoid fluid-pushing.

Fixes #1575
Fixes #2684