Fix loading crash when early loading window is kept minimized\
When the early loading window is minimized, GLFW reports a window size
of 0x0, which gets stored during the control take-over from NeoForge to
Minecraft and causes a crash due to being an invalid window size.

Ensure the stored window size is never below 1x1 when control is passed
between NeoForge and Minecraft.

Fixes #2762